Adjust code to compile on 16-bit platforms (e.g., AVR)
On platforms with 16-bit int, several shift and conversion idioms
relied on int being at least 32 bits:

- Shift constants exceeding 16-bit int need explicit int32_t
  (reduce.h, rounding.h, MLDSA_GAMMA1 in params.h)
- The ghost bound parameter of mld_ntt_butterfly_block reaches up to
  9 * MLD_FQMUL_BOUND and does not fit unsigned int; use uint32_t
- Use unsigned for the Keccak rate in fips202x4.c (as in mlkem-native)
- Keep the nonce computation in mld_yvec_get_poly_lazy unsigned to
  avoid signed/unsigned conversion when uint16_t does not promote
  to int

+++ mldsa/src/reduce.h
@@ -24,7 +24,7 @@
 #define MLD_MONT (-4186625)
 
 /* Upper bound for domain of mld_reduce32() */
-#define MLD_REDUCE32_DOMAIN_MAX (INT32_MAX - (1 << 22))
+#define MLD_REDUCE32_DOMAIN_MAX (INT32_MAX - ((int32_t)1 << 22))
 
 /* Absolute bound for range of mld_reduce32() */
 /* check-magic: 6283009 == (MLD_REDUCE32_DOMAIN_MAX - 255 * MLDSA_Q + 1) */
@@ -114,7 +114,7 @@ __contract__(
 {
   int32_t t;
 
-  t = (a + (1 << 22)) >> 23;
+  t = (a + ((int32_t)1 << 22)) >> 23;
   t = a - t * MLDSA_Q;
   mld_assert((t - a) % MLDSA_Q == 0);
   return t;
